fix: miscellaneous fixes + code cleanup
### Classes
- Add a notice to classes converted from C++
- Add a lightweight fVector3 struct.
- Fix CWheel vector3 fields.

### Feats/Stancer
- Add an option for suspension toe.

### API
- Cleanup and group all utils and extensions.

### CI
- Only trigger translation generator on push to the main branch.
- Don't skip commits in release workflow changelog generation step if the repo has merged PRs.

### Dev/Scripts
- Remove unnecessary string transform in the game data generator script.
