+++ SSV2/includes/classes/gta/rlGamerInfo.lua
@@ -5,6 +5,8 @@
 -- this code as long as you respect these conditions:
 --	* Credit the owner and contributors.
 --	* Provide a copy of or a link to the original license (GPL-3.0 or later); see LICENSE.md or <https://www.gnu.org/licenses/>.
+--
+-- Converted from Yimura's C++ GTA V Classes: https://github.com/Yimura/GTAV-Classes (forked here: https://github.com/Mr-X-GTA/GTAV-Classes-1)
 
 
 local CStructView = require("includes.classes.gta.CStructView")
