+++ SSV2/includes/lib/translations/__hashmap.json
@@ -859,5 +859,6 @@
     "YAV3_SYMBOL_DEFS_HINT": 3101380355,
     "YAV3_LIST_VIEW_HINT": 3504302049,
     "YAV3_VEH_ANIM_HINT": 709579667,
-    "YRV3_HUB_TECH_NOT_ASSIGNED_TT": 1599189488
+    "YRV3_HUB_TECH_NOT_ASSIGNED_TT": 1599189488,
+    "VEH_STANCE_TOE": 834116414
 }
