+++ SSV2/includes/lib/translations/ja-JP.lua
@@ -859,5 +859,6 @@ return {
 	["YAV3_LIST_VIEW_HINT"] = "アニメーションとシナリオのリスト ビューは、[アクション] タブと [コンパニオン] タブでまったく同じです。つまり、一方のタブで使用したフィルターはもう一方のタブにも存在します。ただし、「履歴」タブと「お気に入り」タブには当てはまりません。",
 	["YAV3_SYMBOL_DEFS_HINT"] = "リスト内の一部のアクションの記号が何なのか考えたことはありますか?\n\n- %s: お気に入りとして保存されました。\n- %s: コマンドとして登録されました。\n- %s: ユーザーが生成しました。",
 	["YAV3_VEH_ANIM_HINT"] = "乗り物中心のアニメーションは徒歩では再生できません。仲間に対しても。",
-	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "このハブには技術者が割り当てられていません。"
+	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "このハブには技術者が割り当てられていません。",
+	["VEH_STANCE_TOE"] = "つま先"
 }
