+++ SSV2/includes/lib/translations/ko-KR.lua
@@ -859,5 +859,6 @@ return {
 	["YAV3_LIST_VIEW_HINT"] = "애니메이션 및 시나리오 목록 보기는 '작업' 탭과 '컴패니언' 탭에서 완전히 동일합니다. 즉, 한 탭에서 사용하는 모든 필터가 다른 탭에도 존재한다는 의미입니다. 그러나 '기록' 및 '즐겨찾기' 탭에는 해당되지 않습니다.",
 	["YAV3_SYMBOL_DEFS_HINT"] = "목록에 있는 일부 작업의 기호가 무엇인지 궁금하신가요?\n\n- %s: 즐겨찾기로 저장됨.\n- %s: 명령으로 등록됨.\n- %s: 사용자 생성.",
 	["YAV3_VEH_ANIM_HINT"] = "차량 중심 애니메이션은 도보로 재생할 수 없습니다. 동료들에게도요.",
-	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "이 허브에는 기술자가 할당되지 않습니다."
+	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "이 허브에는 기술자가 할당되지 않습니다.",
+	["VEH_STANCE_TOE"] = "발가락"
 }
