+++ SSV2/includes/lib/translations/ru-RU.lua
@@ -859,5 +859,6 @@ return {
 	["YAV3_SYMBOL_DEFS_HINT"] = "Вы когда-нибудь задумывались, что означают символы некоторых действий в списке?\n\n- %s: Сохранено как избранное.\n- %s: Зарегистрировано как команда.\n- %s: Создано пользователем.",
 	["YAV3_LIST_VIEW_HINT"] = "Представления списка анимации и сценариев одинаковы на вкладках «Действия» и «Сопутствующие», что означает, что любые фильтры, которые вы используете на одной вкладке, будут существовать и на другой. Однако это не относится к вкладкам «история» и «избранное».",
 	["YAV3_VEH_ANIM_HINT"] = "Анимацию, ориентированную на транспортные средства, нельзя воспроизводить пешком; даже для компаньонов.",
-	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "К этому хабу не назначен ни один технический специалист."
+	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "К этому хабу не назначен ни один технический специалист.",
+	["VEH_STANCE_TOE"] = "Палец на ноге"
 }
