+++ SSV2/includes/lib/translations/zh-CN.lua
@@ -859,5 +859,6 @@ return {
 	["YAV3_SYMBOL_DEFS_HINT"] = "有没有想过列表中某些操作上的符号是什么？\n\n- %s：保存为收藏夹。\n- %s：注册为命令。\n- %s：用户生成。",
 	["YAV3_LIST_VIEW_HINT"] = "“操作”选项卡和“同伴”选项卡中的动画和场景列表视图完全相同，这意味着您在一个选项卡中使用的任何过滤器也将存在于另一个选项卡中。然而，“历史记录”和“收藏夹”选项卡的情况并非如此。",
 	["YAV3_VEH_ANIM_HINT"] = "以车辆为中心的动画无法步行播放；即使对于同伴。",
-	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "没有技术人员被分配到该中心。"
+	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "没有技术人员被分配到该中心。",
+	["VEH_STANCE_TOE"] = "脚趾"
 }
