+++ SSV2/includes/lib/translations/zh-TW.lua
@@ -859,5 +859,6 @@ return {
 	["YAV3_LIST_VIEW_HINT"] = "「操作」標籤和「同伴」標籤中的動畫和場景清單視圖完全相同，這表示您在一個標籤中使用的任何篩選器也將存在於另一個標籤中。然而，「歷史記錄」和「收藏夾」標籤的情況並非如此。",
 	["YAV3_SYMBOL_DEFS_HINT"] = "有沒有想過清單中某些操作上的符號是什麼？ \n\n- %s：儲存為收藏夾。 \n- %s：註冊為指令。 \n- %s：使用者生成。",
 	["YAV3_VEH_ANIM_HINT"] = "以車輛為中心的動畫無法步行播放；即使對於同伴。",
-	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "沒有技術人員被分配到該中心。"
+	["YRV3_HUB_TECH_NOT_ASSIGNED_TT"] = "沒有技術人員被分配到該中心。",
+	["VEH_STANCE_TOE"] = "腳趾"
 }
