--- SSV2/includes/lib/utils.lua
@@ -1,1758 +1,0 @@
--- Copyright (C) 2026 SAMURAI (xesdoog) & Contributors.
--- This file is part of Samurai's Scripts.
---
--- Permission is hereby granted to copy, modify, and redistribute
--- this code as long as you respect these conditions:
---	* Credit the owner and contributors.
---	* Provide a copy of or a link to the original license (GPL-3.0 or later); see LICENSE.md or <https://www.gnu.org/licenses/>.
-
-
----@diagnostic disable: lowercase-global
-
-math.randomseed(os.time())
-
-local LUA_TABLE_OVERHEAD <const> = 3 * 0x8 -- 0x18
-local Chrono <const>             = require("includes.modules.Chrono")
-local Cast                       = require("includes.modules.Cast")
-local fMatrix44                  = require("includes.classes.gta.fMatrix44")
-Bit                              = require("includes.modules.Bit")
-
-
-Time      = Chrono.Time
-Timer     = Chrono.Timer
-TimePoint = Chrono.TimePoint
-DateTime  = Chrono.DateTime
-yield     = coroutine.yield
-sleep     = Time.Sleep
-_F        = string.format
-
--- Wrapper for `Translator:Translate`
----@param label string
----@return string
-function _T(label)
-	if (not Translator) then
-		return label
-	end
-
-	return Translator:Translate(label)
-end
-
--- Lua version of Bob Jenskins' "Jenkins One At A Time" hash function
---
--- https://en.wikipedia.org/wiki/Jenkins_hash_function
----@param key string
----@return joaat_t
-local function Joaat(key)
-	local hash = 0
-	key = key:lower()
-
-	for i = 1, #key do
-		hash = hash + string.byte(key, i)
-		hash = hash + (hash << 10)
-		hash = hash & 0xFFFFFFFF
-		hash = hash ~ (hash >> 6)
-	end
-
-	hash = hash + (hash << 3)
-	hash = hash & 0xFFFFFFFF
-	hash = hash ~ (hash >> 11)
-	hash = hash + (hash << 15)
-	hash = hash & 0xFFFFFFFF
-	return hash
-end
-
-_J = _G.joaat or Joaat
-
---#region Global functions
-
--- A dummy function (no-operation).
-function NOP() end
-
--------
-
-local ISINSTANCE_STD_TYPES <const>  = {
-	["table"]    = true,
-	["string"]   = true,
-	["number"]   = true,
-	["boolean"]  = true,
-	["function"] = true,
-	["userdata"] = true
-}
-local ISINSTANCE_MATH_TYPES <const> = {
-	["integer"] = true,
-	["float"]   = true
-}
--- Returns whether `obj` is an instance of `T`. Can be used instead of Lua's default `type` function.
---
--- **Usage Example**:
--- - String types, similar to the default `type` function:
---```Lua
--- print(IsInstance(123, "number")) -> true
---```
--- - Math types, similar to the default `math.type` function:
---```Lua
--- print(IsInstance(123, "float")) -- -> false
--- print(IsInstance(1.23, "float")) -- -> true
---```
--- - Classes:
---```Lua
--- local myveh = LocalPlayer:GetVehicle()
--- print(IsInstance(myveh, Vehicle)) -- -> true
--- print(IsInstance(myveh, Object)) -- -> false
--- print(IsInstance(myveh, Entity)) -- -> true
--- print(IsInstance(myveh, "table")) -- -> false
--- print(IsInstance(myveh, {})) -- -> false
---```
----@param obj any
----@param T anyval
----@return boolean
-function IsInstance(obj, T)
-	if (obj == nil) then
-		return T == "nil"
-	end
-
-	local obj_type <const> = type(obj)
-	local T_type <const>   = type(T)
-	local obj_mt           = getmetatable(obj)
-	local T_mt             = getmetatable(T)
-
-	if (T_type == "string" and T == "pointer" and obj_type == "userdata") then
-		return (obj_mt and type(obj_mt.rip) == "function" or false)
-	end
-
-	if (T_type == "table") then
-		if (obj_type == "userdata" and T.__type == "vec3") then
-			return obj_mt and obj_mt.__type == T.__type
-		end
-
-		local is_obj = rawget(T, "__index") ~= nil
-			or rawget(T, "__type") ~= nil
-			or rawget(T, "__base") ~= nil
-			or T_mt ~= nil
-			or (type(T_mt) == "table" and rawget(T_mt, "__call") ~= nil)
-
-		if (is_obj) then
-			if (obj_type == "table" and obj.__type and obj.__type == T.__type) then
-				return true
-			end
-
-			while obj_mt do
-				if (obj_mt == T) then
-					return true
-				end
-				obj_mt = rawget(obj_mt, "__base")
-			end
-			return false
-		else
-			if (obj_mt or T_mt) then
-				return obj_mt == T_mt
-			end
-		end
-	end
-
-	if (T_type == "string") then
-		if ((obj_type) == "number" and ISINSTANCE_MATH_TYPES[T]) then
-			return math.type(obj) == T
-		end
-
-		if (ISINSTANCE_STD_TYPES[T]) then
-			if (obj_type == "table") then
-				return (obj_type == T and obj_mt == nil)
-			else
-				return obj_type == T
-			end
-		end
-	end
-
-	return false
-end
-
-local SIZEOF_TYPES <const> = {
-	["boolean"]   = 0x1,
-	-- ["DataBuffer"] = function(t) return t:Size() end,
-	["fMatrix44"] = 0x40,
-	["function"]  = 0x8,
-	["nil"]       = 0x0,
-	["number"]    = function(t) return math.sizeof(t) end,
-	["pointer"]   = 0x8,
-	["string"]    = function(t) return string.isempty(t) and 0 or (#t + 1) end,
-	["vec2"]      = 0x8,
-	["vec3"]      = 0xC,
-	["vec4"]      = 0x10,
-}
--- Returns a symbolic size, not actual memory usage.
----@param T any
----@param seen? table Circular reference
----@return number
-function SizeOf(T, seen)
-	local T_type <const> = type(T)
-	local resolved_type -- fwd decl
-
-	if (T_type == "table" or T_type == "userdata") then
-		if (IsInstance(T, "pointer")) then
-			resolved_type = "pointer"
-		elseif (IsInstance(T.__type, "string")) then
-			resolved_type = T.__type
-		end
-	else
-		resolved_type = T_type
-	end
-
-	local known = SIZEOF_TYPES[resolved_type]
-	if (known) then
-		if (type(known) == "function") then
-			return known(T)
-		end
-		return known
-	end
-
-	if (T_type == "table") then
-		local size_method = T.Size
-		if (type(size_method) == "function") then
-			return size_method(T)
-		end
-
-		if (IsInstance(T.m_size, "number")) then
-			return T.m_size
-		end
-
-		if (T.__enum or IsInstance(T.__sizeof, "function")) then
-			return T.__sizeof(T)
-		end
-
-		if (IsInstance(T.__len, "function")) then
-			return T.__len(T)
-		end
-
-		if (IsInstance(T.__type, "string")) then
-			return GenericClass.m_size -- 0x40
-		end
-
-		seen = seen or {}
-		if (seen[T]) then
-			return 0
-		end
-		seen[T] = true
-		return table.sizeof(T, seen)
-	end
-
-	return 0
-end
-
--- A switch-case construct without the closure overhead of `Switch`.
---
--- **Usage Example**:
---
---```Lua
--- local cases = {
---     [1] = function() return "one" end,
---     [2] = function() return "two" end,
---     default = "other"
--- }
--- local result = Match(value, cases)
---```
----@generic R1, R2, R3, R4, R5, R6, R7, R8, R9, R10
----@generic K: anyval
----@generic V
----@param case K
----@param cases table<K, ValueOrFunction<V>>
----@return V|R1, R2?, R3?, R4?, R5?, R6?, R7?, R8?, R9?, R10?
-function Match(case, cases)
-	local result = cases[case] or cases["default"]
-	assert(result ~= nil, "No case matched and no default provided!")
-	return (type(result) == "function") and result() or result
-end
-
--- A switch-case construct. This is only beneficial if you have **A LOT** of `if` statements;
---
--- otherwise regular `if-elseif-else` chains or cached lookup tables are faster and create less overhead.
---
--- **[Note]**: Avoid calling this per-frame. This function creates a closure every time it's called and the reason
---
--- for this is that it allows for more flexible and dynamic case handling. Also looks fancier lol
---
--- Example:
---
---```Lua
--- local result = Switch(value) {
---     [1] = function() return "one" end,
---     [2] = function() return "two" end,
---     default = "default"
--- }
---```
---
--- If you want a version with less overhead, use `Match` instead:
---
---```Lua
--- local cases = {
---     [1] = function() return "one" end,
---     [2] = function() return "two" end,
---     default = "default"
--- }
--- local result = Match(value, cases)
---```
----@param case anyval
-function Switch(case)
-	return function(cases)
-		return Match(case, cases)
-	end
-end
-
----@param e Enum
----@param index integer
-function EnumToString(e, index)
-	if (type(e) ~= "table") then
-		return ""
-	end
-
-	for k, v in pairs(e) do
-		if (v == index) then
-			return tostring(k)
-		end
-	end
-
-	return "Unknown"
-end
-
--- Pauses execution until a condition is met.
---
--- All logic after this call will only execute once the provided function returns a truthy value.
---
--- If the condition isn't met before the optional timeout is reached, `TaskWait` throws an error to prevent further execution.
----@param func fun(args: any): boolean
----@param args any
----@param timeout? milliseconds  -- Optional timeout in milliseconds. Defaults to 1200ms.
-function TaskWait(func, args, timeout)
-	local ftype = type(func)
-	if (ftype ~= "function") then
-		error(_F("Invalid argument #1! Function expected, got %s instead", ftype), 2)
-	end
-
-	args    = args or {}
-	timeout = timeout or 1200
-
-	if (timeout <= 0) then
-		error("Optional timeout must be greater than zero.", 2)
-	end
-
-	-- It is essential to use IsInstance here instead of type(args) == "table"
-	-- otherwise passing objects and userdata will fail.
-	if not IsInstance(args, "table") then
-		args = { args }
-	end
-
-	local start_time = Time.Millis()
-	while (not func(table.unpack(args))) do
-		if ((Time.Millis() - start_time) > timeout) then
-			error("Timeout reached!")
-		end
-		yield()
-	end
-
-	return true
-end
-
----@param t table
----@param mt table|metatable
----@param seen? set<table>
-function RecursiveSetMetatable(t, mt, seen)
-	seen = seen or {}
-	if (seen[t]) then
-		return
-	end
-	seen[t] = true
-
-	for _, v in pairs(t) do
-		if (type(v) == "table" and getmetatable(v) == nil) then
-			RecursiveSetMetatable(v, mt, seen)
-		end
-	end
-
-	return setmetatable(t, mt)
-end
-
--- Simply adds a number suffix if a file with the same name and extension already exists.
----@param base_name string
----@param extension string
----@return string
-function GenerateUniqueFilename(base_name, extension)
-	local filename = string.format("%s%s", base_name, extension)
-	local suffix = 0
-
-	while (io.exists(filename)) do
-		suffix = suffix + 1
-		filename = string.format("%s_%d%s", base_name, suffix, extension)
-	end
-
-	return filename
-end
-
---#endregion
-
---#region extensions
-
---#region Lua API
-
----@param src string Filename with extension
----@param dest string Filename with extension
----@return boolean, any
-function io.copy(src, dest)
-	if (not io.exists(src)) then
-		log.warning("[File I/O]: Copy operation failed. Source file does not exist.")
-		return false, nil
-	end
-
-	local f1 <close> = io.open(src, "r")
-	local f2 <close> = io.open(dest, "w")
-	if (not f1 or not f2) then
-		return false, nil
-	end
-
-	local data = f1:read("a")
-	if (not data) then
-		return false, nil
-	end
-
-	f2:write(data)
-	f2:flush()
-
-	return true, data
-end
-
--- Returns the stat with the appropriate character prefix.
---
--- For online stats, you can pass a stat starting with `MPX`, `MP0`, `MP1`, or `MP_STAT`
---
--- For single player, you can pass a stat starting with `SPX`, `SP0`, `SP1`, or `SP2`
----@param stat_name string
----@return string
-function stats.get_prefixed_stat(stat_name)
-	local char_idx = Game.GetCharacterIndex()
-	if (stat_name:startswith("MPX") or stat_name:startswith("SPX")) then
-		return stat_name:replace_char(3, tostring(char_idx))
-	end
-
-	if (stat_name:startswith("MP_STAT")) then
-		local ret, _ = stat_name:replace("MP_STAT", _F("MP%d", char_idx))
-		return ret
-	end
-
-	return stat_name
-end
-
----@param stat_name string
----@return string
-function stats.get_string(stat_name)
-	stat_name = stats.get_prefixed_stat(stat_name)
-	return STATS.STAT_GET_STRING(_J(stat_name), -1)
-end
-
----@param stat_name string
-function stats.set_string(stat_name, v)
-	stat_name = stats.get_prefixed_stat(stat_name)
-	if (not stat_name:startswith("MP") or not stat_name:startswith("SP")) then
-		return
-	end
-
-	STATS.STAT_SET_STRING(_J(stat_name), v, true)
-end
-
----@param stat_name string
----@param v number
----@param min? number
----@param max? number
-function stats.increment_stat(stat_name, v, min, max)
-	min = min or 0
-	max = max or 100
-
-	local stat_get, stat_set
-	if (math.type(v) == "integer") then
-		stat_get, stat_set = stats.get_int, stats.set_int
-	elseif (math.type(v) == "float") then
-		stat_get, stat_set = stats.get_float, stats.set_float
-	end
-
-	if (not stat_get or not stat_set) then
-		return
-	end
-
-	local sum = stat_get(stat_name) + v
-	if (sum < min or sum > max) then
-		return
-	end
-
-	stat_set(stat_name, sum)
-end
-
-local newptr <const> = memory.pointer.new
----@return pointer
-function memory.pointer:new(addr)
-	if (addr == 0) then return nullptr end
-	return newptr(self, addr)
-end
-
--- Equality comparator for pointer objects.
----@type Comparator<pointer, pointer>
-function memory.pointer:__eq(right)
-	return self:get_address() == right:get_address()
-end
-
--- Allows you to compare pointers to numbers, pointers, or nullptr.
---
--- **Note:** When comparing with a number, the return will indicate whether the pointer's address equals `v`.
----@param v pointer|nullptr|number
----@return boolean
-function memory.pointer:is_equal(v)
-	if (v == self) then
-		return true
-	end
-
-	if (type(v) == "number") then
-		return self:get_address() == v
-	end
-
-	if (not IsInstance(v, "pointer")) then
-		return false
-	end
-
-	return self:get_address() == v:get_address()
-end
-
--- Casts the pointer to an object.
---
--- **IMPORTANT:** You must only cast to objects that take a pointer parameter in their constructors otherwise this method will throw.
---
--- **Example Usage:**
---```lua
---print(memory.handle_to_ptr(self.get_ped()):add(0x10B8):as(CPedWeaponManager))
---```
----@generic T
----@param obj T
----@return T
-function memory.pointer:as(obj)
-	if (self:is_null()) then
-		error("Attempt to cast a null pointer")
-	end
-
-	local obj_type = type(obj)
-	if (obj_type ~= "table") then
-		error(_F("Invalid parameter #1: Table/class expected, got %s instead.", obj_type))
-	end
-
-	if (not obj.__ptr_ctor) then
-		error(_F("Class '%s' constructor does not expect a pointer.", obj.__type or tostring(obj)))
-	end
-
-	if (type(obj.new) == "function") then
-		return obj.new(self)
-	end
-
-	if (type(obj.init) == "function") then
-		return obj:init(self)
-	end
-
-	local mt = getmetatable(obj)
-	if (mt and type(mt.__call) == "function") then
-		return obj(self)
-	end
-
-	error(_F("Class '%s' has no valid pointer constructor", obj.__type or tostring(obj)))
-end
-
--- Retrieves a 32-bit displacement value from the memory address, optionally adding an offset and adjustment.
---
--- **Example Usage:**
--- ```lua
--- displacement = pointer:get_disp32(offset, adjust)
--- ```
----@param offset? integer
----@param adjust? integer
----@return number -- imm32 displacement
-function memory.pointer:get_disp32(offset, adjust)
-	if (self:is_null()) then
-		log.warning("Attempt to get imm32 displacement from a null pointer!")
-		return 0
-	end
-
-	offset = offset or 0
-	adjust = adjust or 0
-	return self:add(offset):get_int() + adjust
-end
-
----@return vec3
-function memory.pointer:get_vec3()
-	if (self:is_null()) then
-		return vec3:zero()
-	end
-
-	local x = self:add(0x4):get_float()
-	local y = self:add(0x8):get_float()
-	local z = self:add(0xC):get_float()
-	return vec3:new(x, y, z)
-end
-
----@param vector3 vec3
-function memory.pointer:set_vec3(vector3)
-	if (self:is_null()) then
-		return
-	end
-
-	self:add(0x4):set_float(vector3.x)
-	self:add(0x8):set_float(vector3.y)
-	self:add(0xC):set_float(vector3.z)
-end
-
----@return vec4
-function memory.pointer:get_vec4()
-	if (self:is_null()) then
-		return vec4:zero()
-	end
-
-	local x = self:add(0x4):get_float()
-	local y = self:add(0x8):get_float()
-	local z = self:add(0xC):get_float()
-	local w = self:add(0x10):get_float()
-
-	return vec4:new(x, y, z, w)
-end
-
----@param vector4 vec4
-function memory.pointer:set_vec4(vector4)
-	if (self:is_null()) then
-		return
-	end
-
-	self:add(0x4):set_float(vector4.x)
-	self:add(0x8):set_float(vector4.y)
-	self:add(0xC):set_float(vector4.z)
-	self:add(0x10):set_float(vector4.w)
-end
-
----@return fMatrix44
-function memory.pointer:get_matrix44()
-	if (self:is_null()) then
-		return fMatrix44:zero()
-	end
-
-	return fMatrix44:new(
-		self:add(0x00):get_float(), self:add(0x04):get_float(), self:add(0x08):get_float(), self:add(0x0C):get_float(),
-		self:add(0x10):get_float(), self:add(0x14):get_float(), self:add(0x18):get_float(), self:add(0x1C):get_float(),
-		self:add(0x20):get_float(), self:add(0x24):get_float(), self:add(0x28):get_float(), self:add(0x2C):get_float(),
-		self:add(0x30):get_float(), self:add(0x34):get_float(), self:add(0x38):get_float(), self:add(0x3C):get_float()
-	)
-end
-
----@param matrix fMatrix44
-function memory.pointer:set_matrix44(matrix)
-	if (self:is_null()) then
-		return
-	end
-
-	local m1 = matrix:R1()
-	local m2 = matrix:R2()
-	local m3 = matrix:R3()
-	local m4 = matrix:R4()
-
-	self:add(0x00):set_float(m1.x); self:add(0x04):set_float(m1.y); self:add(0x08):set_float(m1.z); self:add(0x0C)
-		:set_float(m1.w)
-
-	self:add(0x10):set_float(m2.x); self:add(0x14):set_float(m2.y); self:add(0x18):set_float(m2.z); self:add(0x1C)
-		:set_float(m2.w)
-
-	self:add(0x20):set_float(m3.x); self:add(0x24):set_float(m3.y); self:add(0x28):set_float(m3.z); self:add(0x2C)
-		:set_float(m3.w)
-
-	self:add(0x30):set_float(m4.x); self:add(0x34):set_float(m4.y); self:add(0x38):set_float(m4.z); self:add(0x3C)
-		:set_float(m4.w)
-end
-
----@param size? number bytes
-function memory.pointer:dump(size)
-	if (self:is_null()) then
-		log.debug("Memory Dump<nullptr>")
-		return
-	end
-
-	local result = {}
-	size = size or 0x10
-	for i = 0, size - 1 do
-		local byte = self:add(i):get_byte()
-		table.insert(result, string.format("%02X", byte))
-	end
-
-	log.fdebug(
-		"Memory Dump<0x%X + 0x%X>: %s",
-		self:get_address(),
-		size,
-		table.concat(result, " ")
-	)
-end
-
----@param size? number bytes
-function memory.pointer:create_pattern(size)
-	if (self:is_null()) then
-		return ""
-	end
-
-	local out                      = {}
-	local REG_DIRECT_RANGE <const> = Range(0xC0, 0x100)
-	size                           = size or 0x10
-
-	for i = 0, size - 1 do
-		local byte = self:add(i):get_byte()
-		out[#out + 1] = REG_DIRECT_RANGE:Contains(byte)
-			and "??"
-			or string.format("%02X", byte)
-	end
-
-	return table.concat(out, " ")
-end
-
---#endregion
-
--- stdlib --
-
---#region tablib
-
----@param t table
----@param key string|number
-function table.matchbykey(t, key)
-	if (not t or (table.getlen(t) == 0)) then
-		return
-	end
-
-	for k, v in pairs(t) do
-		if (k == key) then
-			return v
-		end
-	end
-end
-
----@param t table
----@param value any
----@return string|number|nil -- the table key where the value was found or nil
-function table.matchbyvalue(t, value)
-	if not t or (table.getlen(t) == 0) then
-		return nil
-	end
-
-	for k, v in pairs(t) do
-		if v == value then
-			return k
-		end
-	end
-
-	return nil
-end
-
----@param t table
----@param value any
-function table.find(t, value)
-	if (#t == 0) then
-		return false
-	end
-
-	for i = 1, table.getlen(t) do
-		if type(t[i]) == "table" then
-			return table.find(t[i], value)
-		else
-			if type(t[i]) == type(value) then
-				if t[i] == value then
-					return true
-				end
-			end
-		end
-	end
-
-	return false
-end
-
----@generic T
----@param t table<any, T>
----@param pred Predicate<T>
----@return T|nil
-function table.findfirst(t, pred)
-	for _, v in pairs(t) do
-		if pred(v) then
-			return v
-		end
-	end
-
-	return nil
-end
-
----@generic T
----@param arr array<T>
----@param pred Predicate<T>
----@return array<T>
-function table.filter(arr, pred)
-	local out = {}
-
-	for _, v in ipairs(arr) do
-		if pred(v) then
-			out[#out + 1] = v
-		end
-	end
-
-	return out
-end
-
--- Serializes tables in pretty format and avoids circular reference.
----@param tbl table
----@param indent? number
----@param key_order? table
----@param seen? table
-function table.serialize(tbl, indent, key_order, seen)
-	indent = indent or 2
-	seen   = seen or {}
-
-	if (seen[tbl]) then
-		return '"<circular reference>"'
-	end
-
-	seen[tbl] = true
-
-	local function get_indent(level)
-		return string.rep(" ", level)
-	end
-
-	local function serialize_value(v, depth)
-		local __type = type(v)
-		if (__type == "string") then
-			return _F("%q", v)
-		elseif (__type == "number" or __type == "boolean" or __type == "function") then
-			return tostring(v)
-		elseif (__type == "table") then
-			if (table.is_empty(v)) then
-				return "{}"
-			elseif (seen[v]) then
-				return "<circular reference>"
-			else
-				return table.serialize(v, depth, key_order, seen)
-			end
-		elseif (__type == "userdata" or (getmetatable(v) and v.__type)) then
-			if (v.rip and v.get_address) then
-				return _F("<pointer@0x%X>", v:get_address())
-			end
-			return tostring(v)
-		end
-		return "<unsupported>"
-	end
-
-	local is_array, size = table.is_array(tbl)
-	local keys           = {}
-	local pieces         = {}
-	table.insert(pieces, "{\n")
-
-	if (is_array) then
-		for i = 1, size do
-			table.insert(keys, i)
-		end
-	else
-		if (key_order) then
-			for _, k in ipairs(key_order) do
-				if tbl[k] ~= nil then
-					table.insert(keys, k)
-				end
-			end
-
-			for k in pairs(tbl) do
-				if not table.find(keys, k) then
-					table.insert(keys, k)
-				end
-			end
-		else
-			for k in pairs(tbl) do
-				table.insert(keys, k)
-			end
-
-			table.sort(keys, function(a, b)
-				return tostring(a) < tostring(b)
-			end)
-		end
-	end
-
-	for _, k in ipairs(keys) do
-		local v   = tbl[k]
-		local ind = get_indent(indent + 1)
-
-		if (is_array) then
-			table.insert(pieces, ind .. serialize_value(v, indent + 1) .. ",\n")
-		else
-			local key
-			if type(k) == "string" and k:match("^[%a_][%w_]*$") then
-				key = k
-			else
-				key = "[" .. serialize_value(k, indent + 1) .. "]"
-			end
-
-			table.insert(pieces, ind .. key .. " = " .. serialize_value(v, indent + 1) .. ",\n")
-		end
-	end
-
-	table.insert(pieces, get_indent(indent) .. "}")
-	return table.concat(pieces)
-end
-
-function table.print(t)
-	print(table.serialize(t))
-end
-
--- Returns the number of values in a table. Doesn't count nil fields.
----@param t table
----@return number
-function table.getlen(t)
-	if not t then
-		return 0
-	end
-
-	local count = 0
-
-	for _ in pairs(t) do
-		count = count + 1
-	end
-
-	return count
-end
-
----@param t table
----@return boolean, integer size
-function table.is_array(t)
-	local size = #t
-	return size > 0, size
-end
-
----@param t table
----@return boolean
-function table.is_empty(t)
-	return next(t) == nil
-end
-
--- Calculates an estimate of a table's size in memory.
---
--- NOTE: This is **VERY** inaccurate.
----@param t table
----@param seen? table
----@return number
-function table.sizeof(t, seen)
-	if (type(t) ~= "table") then
-		return 0
-	end
-
-	seen = seen or {}
-	if seen[t] then
-		return 0
-	end
-	seen[t] = true
-
-	local size = LUA_TABLE_OVERHEAD
-	for k, v in pairs(t) do
-		size = size + SizeOf(k, seen) + SizeOf(v, seen)
-	end
-
-	return size
-end
-
--- Returns the number of duplicate items in a table.
----@param t table
----@param value string | number | integer | table
-function table.getduplicates(t, value)
-	local count = 0
-	for _, v in ipairs(t) do
-		if (value == v) then
-			count = count + 1
-		end
-	end
-	return count
-end
-
--- Removes duplicate items from a table and returns a new one with the results.
---
--- If `debug` is set to `true`, it adds a table with duplicate items to the return as well.
----@param t table
----@param debug? boolean
-function table.removeduplicates(t, debug)
-	local t_exists, t_clean, t_dupes, t_result = {}, {}, {}, {}
-
-	for _, v in ipairs(t) do
-		if not t_exists[v] then
-			t_clean[#t_clean + 1] = v
-			t_exists[v] = true
-		else
-			if debug then
-				t_dupes[#t_dupes + 1] = v
-			end
-		end
-	end
-
-	if debug then
-		t_result.clean = t_clean
-		t_result.dupes = t_dupes
-	end
-
-	return debug and t_result or t_clean
-end
-
----@generic T : table
----@param t T
----@param seen? table
----@return T
-function table.copy(t, seen)
-	seen = seen or {}
-	if seen[t] then
-		return seen[t]
-	end
-
-	local out = {}
-	seen[t] = out
-
-	for k, v in pairs(t) do
-		if (type(v) == "table") then
-			out[k] = table.copy(v, seen)
-		else
-			out[k] = v
-		end
-	end
-
-	return out
-end
-
--- Merges `src` table into `dest` table.
----@param src table
----@param dest table
----@param path string?
----@param seen table?
-function table.merge(src, dest, path, seen)
-	if (src == dest) then return end
-
-	seen = seen or {}
-	if (seen[src]) then return end
-	seen[src] = true
-
-	for k, v in pairs(src) do
-		if (dest[k] == nil) then
-			dest[k] = v
-		elseif (type(v) == "table" and type(dest[k]) == "table") then
-			local current_path = path and (path .. "." .. tostring(k)) or tostring(k)
-			table.merge(v, dest[k], current_path, seen)
-		end
-	end
-end
-
----@param t1 table
----@param t2 table
-function table.swap(t1, t2)
-	if (t1 == t2) then return end
-
-	local temp = {}
-	for k, v in pairs(t1) do
-		temp[k] = v
-		t1[k]   = nil
-	end
-
-	for k, v in pairs(t2) do
-		t1[k]   = v
-		t2[k]   = temp[k]
-		temp[k] = nil
-	end
-end
-
--- Overwrites `this` table with `src` table.
----@param this table
----@param src table
----@param seen? table
-function table.overwrite(this, src, seen)
-	if (this == src) then return end
-
-	seen = seen or {}
-	if (seen[src]) then return end
-	seen[src] = true
-
-	for k in pairs(this) do
-		if (src[k] == nil) then
-			this[k] = nil
-		end
-	end
-
-	for k, v in pairs(src) do
-		if (type(v) == "table") then
-			if (type(this[k]) ~= "table") then
-				this[k] = {}
-			end
-			table.overwrite(this[k], v, seen)
-		else
-			this[k] = v
-		end
-	end
-end
-
----@param t1 table
----@param t2 table
----@param seen? table
----@return boolean
-function table.is_equal(t1, t2, seen)
-	if (t1 == t2) then
-		return true
-	end
-
-	if (type(t1) ~= type(t2)) then
-		return false
-	end
-
-	if (type(t1) ~= "table") then
-		return false
-	end
-
-	seen = seen or {}
-	if (seen[t1] and seen[t2]) then
-		return true
-	end
-	seen[t1], seen[t2] = true, true
-
-	for k, v in pairs(t1) do
-		if (not table.is_equal(v, t2[k], seen)) then
-			return false
-		end
-	end
-
-	for k in pairs(t2) do
-		if (t1[k] == nil) then
-			return false
-		end
-	end
-
-	return true
-end
-
----@param t table
----@param path string
----@return any
-function table.get_nested_key(t, path)
-	local current = t
-	for key in path:gmatch("[^%.]+") do
-		current = current[key]
-		if (current == nil) then
-			return nil
-		end
-	end
-
-	return current
-end
-
----@param t table
----@param path string
----@param value any
-function table.set_nested_key(t, path, value)
-	local parts = {}
-	for key in path:gmatch("[^%.]+") do
-		parts[#parts + 1] = key
-	end
-
-	local current = t
-	for i = 1, #parts - 1 do
-		local key = parts[i]
-		if (type(current[key]) ~= "table") then
-			current[key] = {}
-		end
-		current = current[key]
-	end
-
-	current[parts[#parts]] = value
-end
-
--- Snapshots a table into a string.
----@generic K, V, P
----@param t table
----@param opts? { out?: table, path?: string, ignored_keys: Set<string> | Predicate<K, V, P?>}
----@return string
-function table.snapshot(t, opts)
-	opts = opts or {}
-	out  = opts.out or {}
-	path = opts.path or ""
-
-	for k, v in pairs(t) do
-		if (opts.ignored_keys) then
-			if (IsInstance(opts.ignored_keys, Set) and opts.ignored_keys:Contains(k)) then
-				goto continue
-			elseif (type(opts.ignored_keys) == "function" and opts.ignored_keys(k, v, path)) then
-				goto continue
-			end
-		end
-
-		local p = path .. "." .. tostring(k)
-		out[#out + 1] = p .. ":" .. type(v)
-		if type(v) == "table" then
-			table.snapshot(v, { out = out, path = p, ignored_keys = opts.ignored_keys })
-		end
-
-		::continue::
-	end
-
-	table.sort(out)
-	return table.concat(out, "|")
-end
-
----@generic K, V
----@param t table<K, V>
----@param pred fun(k: K, v: V): boolean
----@return boolean -- success/failure
-function table.erase_first_if(t, pred)
-	for k, v in pairs(t) do
-		if (pred(k, v)) then
-			t[k] = nil
-			return true
-		end
-	end
-
-	return false
-end
-
----@generic K, V
----@param t table<K, V>
----@param pred fun(k: K, v: V): boolean
----@return boolean, integer -- success/failure and count of removed items
-function table.erase_if(t, pred)
-	local count = 0
-	local is_array, size = table.is_array(t)
-	if (is_array) then
-		for i = size, 1, -1 do
-			if pred(i, t[i]) then
-				table.remove(t, i)
-				count = count + 1
-			end
-		end
-	else
-		local temp = {}
-		for k, v in pairs(t) do
-			if pred(k, v) then
-				temp[#temp + 1] = k
-				count = count + 1
-			end
-		end
-
-		for _, k in ipairs(temp) do
-			t[k] = nil
-		end
-
-		temp = nil
-	end
-
-	return count > 0, count
-end
-
----@param t table
-function table.clear(t)
-	for k in pairs(t) do
-		t[k] = nil
-	end
-end
-
---#endregion
-
---#region stringlib
-
--- Generates a random string.
----@param length? number
----@param isalnum? boolean Alphanumeric
----@return string
-function string.random(length, isalnum)
-	local str_table = {}
-	local charset   = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-	length          = length or math.random(1, 10)
-	length          = math.min(length or math.random(1, 10), 128)
-
-	if (isalnum) then
-		charset = charset .. "0123456789"
-	end
-
-	for _ = 1, length do
-		local index = math.random(1, #charset)
-		table.insert(str_table, charset:sub(index, index))
-	end
-
-	return table.concat(str_table)
-end
-
--- Returns whether a string is alphabetic.
----@param str string
----@return boolean
-function string.isalpha(str)
-	return str:match("^%a+$") ~= nil
-end
-
--- Returns whether a string is numeric.
----@param str string
----@return boolean
-function string.isdigit(str)
-	return str:match("^%d+$") ~= nil
-end
-
--- Returns whether a string is alpha-numeric.
----@param str string
----@return boolean
-function string.isalnum(str)
-	return str:match("^%w+$") ~= nil
-end
-
----@param str string
----@return boolean
-function string.iswhitespace(str)
-	return str:match("^%s*$") ~= nil
-end
-
----@param str? string
----@return boolean
-function string.isnull(str)
-	return str == nil
-end
-
----@param str string
-function string.isempty(str)
-	return #str == 0
-end
-
----@param str string?
----@return boolean
-function string.isnullorempty(str)
-	if str == nil then
-		return true
-	end
-
-	return str:isempty()
-end
-
----@param str string?
----@return boolean
-function string.isnullorwhitespace(str)
-	if str == nil then
-		return true
-	end
-
-	return str:iswhitespace()
-end
-
--- A string is considered invalid if any of these conditions are true:
--- - `nil`
--- - `empty`
--- - `white space`
----@param str? string
----@return boolean
-function string.isvalid(str)
-	return type(str) == "string"
-		and not str:isempty()
-		and not str:iswhitespace()
-end
-
----@param str string
----@return boolean
-function string.is_base64(str)
-	if (not string.isvalid(str)) then
-		return false
-	end
-
-	return (#str % 4 == 0) and (str:match("^[A-Za-z0-9+/]+=?=?$") ~= nil)
-end
-
--- Returns whether a string starts with the provided prefix.
----@param str string
----@param prefix string
----@return boolean
-function string.startswith(str, prefix)
-	return str:sub(1, #prefix) == prefix
-end
-
--- Returns whether a string contains the provided substring.
----@param str string
----@param sub string
----@return boolean
-function string.contains(str, sub)
-	return str:find(sub, 1, true) ~= nil
-end
-
--- Returns whether a string ends with the provided suffix.
----@param str string
----@param suffix string
----@return boolean
-function string.endswith(str, suffix)
-	return str:sub(- #suffix) == suffix
-end
-
--- Inserts a string into another string at the given position.
----@param str string
----@param pos integer
----@param text string
-function string.insert(str, pos, text)
-	pos = math.max(1, math.min(pos, #str + 1))
-	return str:sub(1, pos) .. text .. str:sub(pos)
-end
-
--- Replaces all occurrances of `old` string with `new` string.
---
--- Returns the new string and the count of all occurrances.
----@param str string
----@param old string
----@param new string
----@return string, number
-function string.replace(str, old, new)
-	if old == "" then
-		return str, 0
-	end
-
-	return str:gsub(old:gsub("([^%w])", "%%%1"), new)
-end
-
----@param str string
----@param pos integer
----@param new_cahr string
-function string.replace_char(str, pos, new_cahr)
-	pos = math.max(1, math.min(pos, #str + 1))
-	local prefix = str:sub(1, pos - 1)
-	local suffix = str:sub(pos + 1)
-
-	return _F("%s%s%s", prefix, new_cahr, suffix)
-end
-
--- Joins a table of strings using a separator.
----@param sep string
----@param tbl string[]
----@return string
-function string.join(sep, tbl)
-	return table.concat(tbl, sep)
-end
-
--- Removes leading and trailing white space from a string.
----@param str string
----@return string
-function string.trim(str)
-	return str:match("^%s*(.-)%s*$")
-end
-
----@param str string
----@param width integer
-function string.wrap(str, width)
-	local out = {}
-	for i = 1, #str, width do
-		table.insert(out, str:sub(i, i + width - 1))
-	end
-	return table.concat(out, "\n")
-end
-
--- Splits a string by a separator and returns a table of strings.
----@param str string
----@param sep string
----@param maxsplit? integer Optional: limit the number of splits.
----@return string[]
-function string.split(str, sep, maxsplit)
-	local result, count = {}, 0
-	local pattern = "([^" .. sep .. "]+)"
-
-	for part in str:gmatch(pattern) do
-		table.insert(result, part)
-		count = count + 1
-
-		if maxsplit and count >= maxsplit then
-			local rest = str:match("^" .. (("([^" .. sep .. "]+)" .. sep):rep(count)) .. "(.+)$")
-			if rest then
-				table.insert(result, rest)
-			end
-			break
-		end
-	end
-
-	return result
-end
-
--- Same as `string.split` but starts from the right.
----@param str string
----@param sep string
----@param maxsplit? integer Optional: limit the number of splits.
----@return string[]
-function string.rsplit(str, sep, maxsplit)
-	local splits = {}
-
-	for part in string.gmatch(str, "([^" .. sep .. "]+)") do
-		table.insert(splits, part)
-	end
-
-	local total = #splits
-	if not maxsplit or maxsplit <= 0 or maxsplit >= total - 1 then
-		return splits
-	end
-
-	local head = {}
-	for i = 1, total - maxsplit - 1 do
-		table.insert(head, splits[i])
-	end
-
-	local tail = table.concat(splits, sep, total - maxsplit, total)
-	table.insert(head, tail)
-	return head
-end
-
--- Python-like `partition` implementation: Splits a string into 3 parts: before, separator, after
----@param str string
----@param sep string
----@return string, string, string
-function string.partition(str, sep)
-	local start_pos, end_pos = str:find(sep, 1, true)
-
-	if not start_pos then
-		return str, "", ""
-	end
-
-	return str:sub(1, start_pos - 1), sep, str:sub(end_pos + 1)
-end
-
--- Same as `string.partition` but starts from the right.
----@param str string
----@param sep string
----@return string, string, string
-function string.rpartition(str, sep)
-	local start_pos, end_pos = str:reverse():find(sep:reverse(), 1, true)
-
-	if not start_pos then
-		return "", "", str
-	end
-
-	local rev_index = #str - end_pos + 1
-	return str:sub(1, rev_index - 1), sep, str:sub(rev_index + #sep)
-end
-
----@param str string
----@param len number
----@param char string
----@return string
-function string.padleft(str, len, char)
-	return _F("%s%s", string.rep(char or " ", math.max(0, len - #str)), str)
-end
-
----@param str string
----@param len number
----@param char string
----@return string
-function string.padright(str, len, char)
-	return _F("%s%s", str, string.rep(char or " ", math.max(0, len - #str)))
-end
-
--- Capitalizes the first letter in a string.
----@param str string
----@return string
-function string.capitalize(str)
-	return (str:lower():gsub("^%l", string.upper))
-end
-
--- Capitalizes the first letter of each word in a string.
----@param str string
----@return string
-function string.titlecase(str)
-	return (str:gsub("(%a)([%w_']*)", function(a, b)
-		return a:upper() .. b:lower()
-	end))
-end
-
----@return string
-function string.pascalcase(str)
-	return str:lower():gsub("%_+", " "):gsub("%-+", ""):titlecase():gsub("%s+", "")
-end
-
----@return string
-function string.snakecase(str)
-	return str:lower():gsub("[%s%-]", "_")
-end
-
----@param value number|string
----@return string
-function string.formatint(value)
-	local s, _ = tostring(value):reverse():gsub("%d%d%d", "%1,"):reverse():gsub("^,", "")
-	return s
-end
-
----@param value number|string
----@param currency? string
----@return string
-function string.formatmoney(value, currency)
-	return _F("%s%s", currency or "$", string.formatint(value))
-end
-
----@param str string
-function string.hex2string(str)
-	return (str:gsub("%x%x", function(digits)
-		return string.char(tonumber(digits, 16))
-	end))
-end
-
----@param v string|number
----@return string
-function string.hex(v)
-	local _type = type(v)
-	if (_type == "string") then
-		return (string.gsub(v, ".", function(char)
-			return _F("%02x", string.byte(char))
-		end))
-	elseif (_type == "number") then
-		return _F("0x%X", v)
-	end
-
-	return ""
-end
-
---#endregion
-
---#region mathlib
-
----@param n number
----@param x number
----@return number
-function math.round(n, x)
-	return tonumber(string.format("%." .. (x or 0) .. "f", n)) or 0
-end
-
----@param ... any Varargs or array of numbers.
----@return number
-function math.sum(...)
-	local result = 0
-	local args   = type(...) == "table" and ... or { ... }
-	local __len  = #args
-	if (__len == 0) then
-		return 0
-	end
-
-	for i = 1, __len do
-		if (type(args[i]) == "number") then
-			result = result + args[i]
-		end
-	end
-
-	return result
-end
-
----@param n number|nil
----@return boolean
-function math.is_null(n)
-	return type(n) ~= "number" or n == 0
-end
-
----@param n number
----@param min number
----@param max number
----@return boolean
-function math.is_inrange(n, min, max)
-	return n >= min and n <= max
-end
-
--- This ignores floating point precision.
---
--- For normal numbers, use regular equality comparison.
---
--- https://www.lua.org/pil/2.3.html
----@param a float
----@param b float
----@param e? float Optional epsilon (threshold)
----@return boolean
-function math.is_equal(a, b, e)
-	e = e or 1e-6
-	return a == b or math.abs(a - b) < e
-end
-
-local INT_SIZES <const>     = {
-	int8_t   = 0x1,
-	int16_t  = 0x2,
-	int32_t  = 0x4,
-	int64_t  = 0x8,
-	uint8_t  = 0x1,
-	uint16_t = 0x2,
-	uint32_t = 0x4,
-	uint64_t = 0x8,
-}
-local INT_INFERENCE <const> = {
-	["unsigned"] = {
-		{ Cast.AsUint8,  "uint8_t" },
-		{ Cast.AsUint16, "uint16_t" },
-		{ Cast.AsUint32, "uint32_t" },
-		{ Cast.AsUint64, "uint64_t" },
-	},
-	["signed"] = {
-		{ Cast.AsInt8,  "int8_t" },
-		{ Cast.AsInt16, "int16_t" },
-		{ Cast.AsInt32, "int32_t" },
-		{ Cast.AsInt64, "int64_t" },
-	}
-}
----@param n integer
-function math.sizeof(n)
-	local t_n = type(n)
-	assert(t_n == "number",
-		_F("Attempt to call math.sizeof on an invalid value. Number expected, got %s instead!", t_n)
-	)
-
-	if (n == 0 or IsInstance(n, "float")) then
-		return 0x4
-	end
-
-	local c   = Cast(n)
-	local key = n < 0 and "signed" or "unsigned"
-	local _t  = INT_INFERENCE[key]
-
-	for i = 1, #_t do
-		local method, size_key = _t[i][1], _t[i][2]
-		local value = method(c)
-		if (math.is_equal(n, value, 1e-9)) then
-			return INT_SIZES[size_key]
-		end
-	end
-
-	return 0x4
-end
-
----@param v number
----@param min number
----@param max number
----@return number
-function math.clamp(v, min, max)
-	return math.max(min, math.min(max, v))
-end
-
----@param a number
----@param b number
----@param t number delta
-function math.lerp(a, b, t)
-	return a + (b - a) * math.clamp(t, 0, 1)
-end
-
--- Generates a triangular wave oscillating between 1 and -1
----@param t number
----@return number
-function math.tent(t)
-	return 2 * math.abs(2 * (t - math.floor(t + 0.5))) - 1
-end
-
--- 3x² - 2x²
----@param x number
----@return number
-function math.smooth_step(x)
-	return x * x * (3 - 2 * x)
-end
-
----@return -2147483648
-function math.int32_min()
-	return -2147483648
-end
-
----@return 2147483647
-function math.int32_max()
-	return 2147483647
-end
-
----@return 0
-function math.uint32_min()
-	return 0
-end
-
----@return 4294967295
-function math.uint32_max()
-	return 4294967295
-end
-
---#endregion
-
---#endregion
